:hammer: update data type

+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/OneScanActivity.java
@@ -21,7 +21,7 @@ import static com.vijay.jsonwizard.constants.JsonFormConstants.BARCODE_CONSTANTS
 
 public class OneScanActivity extends AppCompatActivity {
 
-    private final int finishDelayed = 4500;
+    private final long finishDelayed = 4500;
     private OneScanHelper oneScanHelper;
 
     @Override
